Update README.md
left a record comments to explain how i stated this project with a little hope that I would feel fun when I see it in the future. 
I hope the day I see it in the future, my developing skill would be better than now. 
upsp , I said a  stupid word.